+++ demo/src/app/components/+dropdown/demos/config/config.ts
@@ -1,16 +1,10 @@
 import { Component } from '@angular/core';
-import { DropdownConfig } from 'ng2-bootstrap/dropdown';
-
-// such override allows to keep some initial values
-
-export function getDropdownConfig(): DropdownConfig {
-  return Object.assign(new DropdownConfig(), {autoClose: 'disabled', keyboardNav: true});
-}
+import { BsDropdownConfig } from 'ngx-bootstrap/dropdown';
 
 @Component({
   selector: 'demo-dropdown-config',
   templateUrl: './config.html',
-  providers: [{provide: DropdownConfig, useFactory: getDropdownConfig}]
+  providers: [{provide: BsDropdownConfig, useValue: {autoClose: false}}]
 })
 export class DemoDropdownConfigComponent {
 }