+++ demo/src/app/components/+dropdown/demos/index.ts
@@ -1,4 +1,3 @@
-import { DropdownDemoComponent } from './dropdown-demo.component';
 import { DemoDropdownBasicComponent } from './basic/basic';
 import { DemoDropdownSplitComponent } from './split/split';
 import { DemoDropdownContainerComponent } from './container/container';
@@ -9,11 +8,13 @@ import { DemoDropupComponent } from './dropup/dropup';
 import { DemoDropdownTriggersManualComponent } from './triggers-manual/triggers-manual';
 import { DemoDropdownDisabledComponent } from './disabled-menu/disabled-menu';
 import { DemoDropdownAlignmentComponent } from './alignment/menu-alignment';
+import { DemoNestedDropdownsComponent } from './nested-dropdowns/nested-dropdowns';
 
 export const DEMO_COMPONENTS = [
-  DropdownDemoComponent, DemoDropdownBasicComponent, DemoDropdownBasicLinkComponent, DemoDropdownSplitComponent,
+  DemoDropdownBasicComponent, DemoDropdownBasicLinkComponent, DemoDropdownSplitComponent,
   DemoDropdownContainerComponent, DemoDropdownKeyboardComponent, DemoDropdownConfigComponent, DemoDropupComponent,
-  DemoDropdownTriggersManualComponent, DemoDropdownDisabledComponent, DemoDropdownAlignmentComponent
+  DemoDropdownTriggersManualComponent, DemoDropdownDisabledComponent, DemoDropdownAlignmentComponent,
+  DemoNestedDropdownsComponent
 ];
 
 export const DEMOS = {
@@ -45,6 +46,10 @@ export const DEMOS = {
   alignment: {
     component: require('!!raw-loader?lang=typescript!./alignment/menu-alignment.ts'),
     html: require('!!raw-loader?lang=markup!./alignment/menu-alignment.html')
+  },
+  nested: {
+    component: require('!!raw-loader?lang=typescript!./nested-dropdowns/nested-dropdowns.ts'),
+    html: require('!!raw-loader?lang=markup!./nested-dropdowns/nested-dropdowns.html')
   },
   keyboard: {
     component: require('!!raw-loader?lang=typescript!./keyboard/keyboard.ts'),