+++ demo/src/app/components/+dropdown/dropdown-section.component.ts
@@ -20,6 +20,7 @@ let titleDoc = require('html-loader!markdown-loader!./docs/title.md');
         <li><a routerLink="." fragment="triggers-manual">Manual triggers</a></li>
         <li><a routerLink="." fragment="disabled-menu">Disabled menu</a></li>
         <li><a routerLink="." fragment="alignment">Menu alignment</a></li>
+        <li><a routerLink="." fragment="nested">Nested dropdowns (experimental)</a></li>
         <li><a routerLink="." fragment="container-body">Append to body</a></li>
         <li><a routerLink="." fragment="keyboard">Single button with keyboard nav</a></li>
         <li><a routerLink="." fragment="dropup">Dropup variation</a></li>
@@ -87,16 +88,23 @@ let titleDoc = require('html-loader!markdown-loader!./docs/title.md');
     <demo-dropdown-alignment></demo-dropdown-alignment>
    </ng-sample-box>
   
+  <h2 routerLink="." fragment="nested" id="nested">Nested dropdowns (experimental)</h2>
+   
+   <ng-sample-box [ts]="demos.nested.component" [html]="demos.nested.html">
+     <demo-nested-dropdowns></demo-nested-dropdowns>
+   </ng-sample-box>
+  
    <h2 routerLink="." fragment="container-body" id="container-body">Append to body</h2>
    
    <ng-sample-box [ts]="demos.container.component" [html]="demos.container.html">
     <demo-dropdown-container></demo-dropdown-container>
   </ng-sample-box>
-  
-  <h2 routerLink="." fragment="keyboard" id="keyboard">Single button with keyboard nav</h2>
-   <ng-sample-box [ts]="demos.keyboard.component" [html]="demos.keyboard.html">
-    <demo-dropdown-keyboard></demo-dropdown-keyboard>
-  </ng-sample-box>
+
+  <!-- not availavle in bs-dropdown version -->
+  <!--<h2 routerLink="." fragment="keyboard" id="keyboard">Single button with keyboard nav</h2>-->
+   <!--<ng-sample-box [ts]="demos.keyboard.component" [html]="demos.keyboard.html">-->
+    <!--<demo-dropdown-keyboard></demo-dropdown-keyboard>-->
+  <!--</ng-sample-box>-->
     
   <h2 routerLink="." fragment="dropup" id="dropup">Dropup variation</h2>
   <p>Trigger dropdown menus above elements by adding <code>.dropup</code> to the parent element.</p>
@@ -111,14 +119,15 @@ let titleDoc = require('html-loader!markdown-loader!./docs/title.md');
   </ng-sample-box>
   
   <h2 routerLink="." fragment="api-reference" id="api-reference">API Reference</h2>
-  <ng-api-doc id="dropdown-directive" directive="DropdownDirective"></ng-api-doc>
-  <ng-api-doc id="dropdown-toggle-directive" directive="DropdownToggleDirective"></ng-api-doc>
-  <ng-api-doc-config id="dropdown-config" type="DropdownConfig"></ng-api-doc-config>
+  <ng-api-doc id="dropdown-directive" directive="BsDropdownDirective"></ng-api-doc>
+  <ng-api-doc id="dropdown-menu-directive" directive="BsDropdownMenuDirective"></ng-api-doc>
+  <ng-api-doc id="dropdown-toggle-directive" directive="BsDropdownToggleDirective"></ng-api-doc>
+  <ng-api-doc-config id="dropdown-config" type="BsDropdownConfig"></ng-api-doc-config>
 </demo-section>`
 })
 export class DropdownSectionComponent {
   public name: string = 'Dropdowns';
-  public src: string = 'https://github.com/valor-software/ng2-bootstrap/tree/development/src/dropdown';
+  public src: string = 'https://github.com/valor-software/ngx-bootstrap/tree/development/src/dropdown';
   public demos: any = DEMOS;
   public titleDoc: string = titleDoc;
 }