+++ demo/src/app/components/+typeahead/typeahead-section.component.ts
@@ -70,7 +70,7 @@ let titleDoc = require('html-loader!markdown-loader!./docs/title.md');
 })
 export class TypeaheadSectionComponent {
   public name:string = 'Typeahead';
-  public src:string = 'https://github.com/valor-software/ng2-bootstrap/tree/development/src/typeahead';
+  public src:string = 'https://github.com/valor-software/ngx-bootstrap/tree/development/src/typeahead';
   public demos: any = DEMOS;
   public titleDoc:string = titleDoc;
 }