+++ src/typeahead/typeahead-container.component.ts
@@ -2,9 +2,9 @@ import {
   Component, ElementRef, TemplateRef, ViewEncapsulation, HostListener
 } from '@angular/core';
 import { isBs3 } from '../utils/ng2-bootstrap-config';
-import { TypeaheadUtils } from './typeahead-utils';
 import { TypeaheadDirective } from './typeahead.directive';
 import { TypeaheadMatch } from './typeahead-match.class';
+import { latinize } from './typeahead-utils';
 
 @Component({
   selector: 'typeahead-container',
@@ -143,7 +143,7 @@ export class TypeaheadContainerComponent {
   public hightlight(match: TypeaheadMatch, query: any): string {
     let itemStr: string = match.value;
     let itemStrHelper: string = (this.parent && this.parent.typeaheadLatinize
-      ? TypeaheadUtils.latinize(itemStr)
+      ? latinize(itemStr)
       : itemStr).toLowerCase();
     let startIdx: number;
     let tokenLen: number;